requirements.txt
Pillow==10.4.0
speedtest==0.0.1
speedtest_cli==2.1.3
